把VideoPlayer掰成了三瓣
